+++ turbopack/crates/turbo-tasks-backend/fuzz/.gitignore
@@ -1,2 +1,3 @@
 artifacts
 corpus
+afl-out
